Move constructor tests to deprecation test module
