+++ tests/test_targets.py
@@ -1,7 +1,5 @@
 """Target tests."""
 
-from itertools import pairwise
-
 import numpy as np
 import pandas as pd
 import pytest
@@ -10,12 +8,6 @@ from pandas.testing import assert_series_equal
 from pytest import param
 
 from baybe.exceptions import IncompatibilityError
-from baybe.targets._deprecated import (
-    LegacyTarget,
-    bell_transform,
-    linear_transform,
-    triangular_transform,
-)
 from baybe.targets.numerical import NumericalTarget as ModernTarget
 from baybe.transformations import (
     AbsoluteTransformation,
@@ -42,229 +34,6 @@ def sample_input() -> pd.Series:
 @pytest.fixture
 def series() -> pd.Series:
     return sample_input()
-
-
-@pytest.mark.parametrize("mode", ["MAX", "MIN"])
-def test_constructor_equivalence_min_max(mode):
-    """
-    Calling the new target class with legacy arguments yields an object equivalent
-    to the legacy object.
-    """  # noqa
-    groups = [
-        (
-            # ------------
-            # Legacy style
-            ModernTarget("t", mode),
-            ModernTarget("t", mode=mode),
-            ModernTarget(name="t", mode=mode),
-            # ------------
-            # Modern style
-            ModernTarget("t", minimize=mode == "MIN"),
-        ),
-        (
-            # ------------
-            # Legacy style
-            ModernTarget("t", mode, (1, 2)),
-            ModernTarget("t", mode, bounds=(1, 2)),
-            ModernTarget("t", mode=mode, bounds=(1, 2)),
-            ModernTarget(name="t", mode=mode, bounds=(1, 2)),
-            ModernTarget("t", mode, (1, 2), "LINEAR"),
-            ModernTarget("t", mode, (1, 2), transformation="LINEAR"),
-            ModernTarget("t", mode, bounds=(1, 2), transformation="LINEAR"),
-            ModernTarget("t", mode=mode, bounds=(1, 2), transformation="LINEAR"),
-            ModernTarget(name="t", mode=mode, bounds=(1, 2), transformation="LINEAR"),
-            # ------------
-            # Modern style
-            ModernTarget.normalize_ramp(
-                name="t", cutoffs=(1, 2), descending=mode == "MIN"
-            ),
-        ),
-    ]
-    for targets in groups:
-        for t1, t2 in pairwise(targets):
-            assert t1 == t2
-
-
-@pytest.mark.parametrize("transformation", ["TRIANGULAR", "BELL"])
-def test_constructor_equivalence_match(transformation):
-    """
-    Calling the new target class with legacy arguments yields an object equivalent
-    to the legacy object.
-    """  # noqa
-    # ------------
-    # Legacy style
-    targets = (
-        ModernTarget("t", "MATCH", (1, 2), transformation),
-        ModernTarget("t", "MATCH", (1, 2), transformation=transformation),
-        ModernTarget("t", "MATCH", bounds=(1, 2), transformation=transformation),
-        ModernTarget("t", mode="MATCH", bounds=(1, 2), transformation=transformation),
-        ModernTarget(
-            name="t", mode="MATCH", bounds=(1, 2), transformation=transformation
-        ),
-    )
-    if transformation == "TRIANGULAR":
-        targets += (
-            ModernTarget("t", "MATCH", (1, 2)),
-            ModernTarget("t", "MATCH", bounds=(1, 2)),
-            ModernTarget("t", mode="MATCH", bounds=(1, 2)),
-            ModernTarget(name="t", mode="MATCH", bounds=(1, 2)),
-        )
-
-    # ------------
-    # Modern style
-    if transformation == "BELL":
-        targets += (ModernTarget.match_bell("t", match_value=1.5, sigma=0.5),)
-    else:
-        targets += (
-            ModernTarget.match_triangular("t", cutoffs=(1, 2)),
-            ModernTarget.match_triangular("t", match_value=1.5, width=1),
-            ModernTarget.match_triangular("t", match_value=1.5, margins=(0.5, 0.5)),
-        )
-
-    for t1, t2 in pairwise(targets):
-        assert t1 == t2
-
-
-@pytest.mark.parametrize(
-    ("legacy", "deprecation", "modern", "expected"),
-    [
-        param(
-            LegacyTarget("t", "MAX"),
-            ModernTarget("t", "MAX"),
-            ModernTarget("t"),
-            sample_input(),
-            id="max",
-        ),
-        param(
-            LegacyTarget("t", "MAX", (0, 1), "LINEAR"),
-            ModernTarget("t", "MAX", (0, 1), "LINEAR"),
-            ModernTarget("t").clamp(0, 1),
-            linear_transform(sample_input(), 0, 1, descending=False),
-            id="max_clamped",
-        ),
-        param(
-            LegacyTarget("t", "MAX", (2, 5), "LINEAR"),
-            ModernTarget("t", "MAX", (2, 5), "LINEAR"),
-            ModernTarget.normalize_ramp("t", (2, 5)),
-            linear_transform(sample_input(), 2, 5, descending=False),
-            id="max_shifted_clamped",
-        ),
-        param(
-            # NOTE: Minimization transformation without bounds is not possible with
-            #   legacy interface."
-            None,
-            None,
-            ModernTarget("t", AffineTransformation(factor=-1)),
-            -sample_input(),
-            id="min_no_bounds",
-        ),
-        param(
-            # NOTE: Minimization transformation without bounds is not possible with
-            #   legacy interface."
-            None,
-            None,
-            ModernTarget("t", minimize=True),
-            -sample_input(),
-            id="min_no_bounds_with_flag",
-        ),
-        param(
-            # NOTE: Minimization without bounds had no effect on the transformation
-            #   of the legacy target since minimization was handled in the construction
-            #   of the acquisition function:
-            #   * https://github.com/emdgroup/baybe/pull/462
-            #   * https://github.com/emdgroup/baybe/issues/460
-            None,  # should be `LegacyTarget("t", "MIN")` but see explanation above
-            ModernTarget("t", "MIN"),
-            ModernTarget("t", minimize=True),
-            -sample_input(),
-            id="min",
-        ),
-        param(
-            LegacyTarget("t", "MIN", (0, 1), "LINEAR"),
-            ModernTarget("t", "MIN", (0, 1), "LINEAR"),
-            ModernTarget.normalize_ramp("t", (0, 1), descending=True),
-            linear_transform(sample_input(), 0, 1, descending=True),
-            id="min_clamped",
-        ),
-        param(
-            LegacyTarget("t", "MIN", (2, 5), "LINEAR"),
-            ModernTarget("t", "MIN", (2, 5), "LINEAR"),
-            ModernTarget.normalize_ramp("t", (2, 5), descending=True),
-            linear_transform(sample_input(), 2, 5, descending=True),
-            id="min_shifted_clamped",
-        ),
-        param(
-            LegacyTarget("t", "MATCH", (-1, 1), "BELL"),
-            ModernTarget("t", "MATCH", (-1, 1), "BELL"),
-            ModernTarget.match_bell("t", match_value=0, sigma=1),
-            bell_transform(sample_input(), -1, 1),
-            id="match_bell_unit_centered",
-        ),
-        param(
-            LegacyTarget("t", "MATCH", (1, 3), "BELL"),
-            ModernTarget("t", "MATCH", (1, 3), "BELL"),
-            ModernTarget.match_bell("t", match_value=2, sigma=1),
-            bell_transform(sample_input(), 1, 3),
-            id="match_bell_unit_shifted",
-        ),
-        param(
-            LegacyTarget("t", "MATCH", (-5, 5), "BELL"),
-            ModernTarget("t", "MATCH", (-5, 5), "BELL"),
-            ModernTarget.match_bell("t", match_value=0, sigma=5),
-            bell_transform(sample_input(), -5, 5),
-            id="match_bell_scaled_centered",
-        ),
-        param(
-            LegacyTarget("t", "MATCH", (2, 6), "BELL"),
-            ModernTarget("t", "MATCH", (2, 6), "BELL"),
-            ModernTarget.match_bell("t", match_value=4, sigma=2),
-            bell_transform(sample_input(), 2, 6),
-            id="match_bell_scaled_shifted",
-        ),
-        param(
-            LegacyTarget("t", "MATCH", (-1, 1), "TRIANGULAR"),
-            ModernTarget("t", "MATCH", (-1, 1), "TRIANGULAR"),
-            ModernTarget.match_triangular("t", cutoffs=(-1, 1)),
-            triangular_transform(sample_input(), -1, 1),
-            id="match_triangular_unit_centered",
-        ),
-        param(
-            LegacyTarget("t", "MATCH", (1, 3), "TRIANGULAR"),
-            ModernTarget("t", "MATCH", (1, 3), "TRIANGULAR"),
-            ModernTarget.match_triangular("t", cutoffs=(1, 3)),
-            triangular_transform(sample_input(), 1, 3),
-            id="match_triangular_unit_shifted",
-        ),
-        param(
-            LegacyTarget("t", "MATCH", (-5, 5), "TRIANGULAR"),
-            ModernTarget("t", "MATCH", (-5, 5), "TRIANGULAR"),
-            ModernTarget.match_triangular("t", cutoffs=(-5, 5)),
-            triangular_transform(sample_input(), -5, 5),
-            id="match_triangular_scaled_centered",
-        ),
-        param(
-            LegacyTarget("t", "MATCH", (2, 6), "TRIANGULAR"),
-            ModernTarget("t", "MATCH", (2, 6), "TRIANGULAR"),
-            ModernTarget.match_triangular("t", cutoffs=(2, 6)),
-            triangular_transform(sample_input(), 2, 6),
-            id="match_triangular_scaled_shifted",
-        ),
-    ],
-)
-def test_target_transformation(
-    series,
-    legacy: LegacyTarget,
-    deprecation: ModernTarget,
-    modern: ModernTarget,
-    expected,
-):
-    """The legacy and modern target variants transform equally."""
-    expected = pd.Series(expected)
-    if legacy is not None:
-        assert_series_equal(legacy.transform(series), expected)
-    if deprecation is not None:
-        assert_series_equal(deprecation.transform(series), expected)
-    assert_series_equal(modern.transform(series), expected)
 
 
 @pytest.mark.parametrize("chained_first", [True, False])
